debug TimeHF cycle

+++ src/EpicDetector.cxx
@@ -455,7 +455,7 @@ void EpicDetector::BuildRawEvent(const std::string &daq,
     if (label == "HF") {
       faster_data_load(data, &hf_data);
       m_TimeHF_prev    = m_TimeHF_current;
-      m_TimeHF_current = (double)timestamp + (double)(qdc_conv_dt_ns(hf_data.tdc));
+      m_TimeHF_current = (double)(timestamp + (long double)(qdc_conv_dt_ns(hf_data.tdc)));
       cout << setprecision(25) << " --> HF data: t_hf = " << m_TimeHF_current << "(DELTA = " << m_TimeHF_current - m_TimeHF_prev << ")" << endl;
       m_RawData->SetTimePrevHF(m_TimeHF_prev);
       m_RawData->SetTimeHF(m_TimeHF_current);
